refactor: detach the `graphql` folder
